+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - added `spawnEndPortal` and `spawnNetherPortal` helper methods to particle state events
 - fixed non Oritech recipes being completely removed ([#1](https://github.com/AlmostReliable/kubejs-oritech/issues/1))
+- fixed duplicate recipe ID check for `deepDrillRegistration` event
 - changed `deepDrillRegistration` event to use proper KubeJS exceptions to display exact source line in case of errors
 
 ## [0.3.0] - 2026-02-06